+++ app/models/foi_attachment.rb
@@ -1,5 +1,5 @@
 # == Schema Information
-# Schema version: 20250304205550
+# Schema version: 20250408105243
 #
 # Table name: foi_attachments
 #
@@ -18,6 +18,8 @@
 #  prominence_reason     :text
 #  masked_at             :datetime
 #  locked                :boolean          default(FALSE)
+#  replaced_at           :datetime
+#  replaced_reason       :string
 #
 
 # models/foi_attachment.rb:
@@ -36,6 +38,10 @@ class FoiAttachment < ApplicationRecord
 
   MissingAttachment = Class.new(StandardError)
 
+  attribute :replacement_file
+  attribute :replacement_body, :string
+  attribute :replaced_filename, :string
+
   belongs_to :incoming_message, inverse_of: :foi_attachments, optional: true
   has_one :info_request, through: :incoming_message, source: :info_request
   has_one :raw_email, through: :incoming_message, source: :raw_email
@@ -45,8 +51,13 @@ class FoiAttachment < ApplicationRecord
   validates_presence_of :content_type
   validates_presence_of :filename
   validates_presence_of :display_size
+  validates :replaced_filename, absence: true, unless: :replacing_or_replaced?
+  validates :replaced_reason, absence: true, unless: :replacing_or_replaced?
+  validates :replaced_reason, presence: true, if: :replacing_or_replaced?
 
   before_validation :ensure_filename!, only: [:filename]
+  before_save :handle_locked
+  before_save :handle_replacements
   before_destroy :delete_cached_file!
 
   scope :binary, -> { where.not(content_type: AlaveteliTextMasker::TextMask) }
@@ -117,10 +128,11 @@ class FoiAttachment < ApplicationRecord
     return @cached_body if @cached_body
 
     begin
-      return file.download if masked?
-    rescue ActiveStorage::FileNotFoundError
+      return file.download if locked? || masked?
+    rescue ActiveStorage::FileNotFoundError => ex
       # file isn't in storage and has gone missing, rescue to allow the masking
       # job to run and rebuild the stored file or even the whole attachment.
+      raise ex if locked?
     end
 
     if persisted?
@@ -146,28 +158,7 @@ class FoiAttachment < ApplicationRecord
   # return the body as it is in the raw email, unmasked without censor rules
   # applied
   def unmasked_body
-    MailHandler.attachment_body_for_hexdigest(
-      raw_email.mail,
-      hexdigest: hexdigest
-    )
-
-  rescue MailHandler::MismatchedAttachmentHexdigest
-    begin
-      attributes = MailHandler.attempt_to_find_original_attachment_attributes(
-        raw_email.mail,
-        body: file.download
-      ) if file.attached?
-
-    rescue ActiveStorage::FileNotFoundError
-      raise MissingAttachment, "attachment missing from storage (ID=#{id})"
-    end
-
-    unless attributes
-      raise MissingAttachment, "attachment missing in raw email (ID=#{id})"
-    end
-
-    update(hexdigest: attributes[:hexdigest])
-    attributes[:body]
+    mail_attributes[:body]
   end
 
   def masked?
@@ -202,7 +193,7 @@ class FoiAttachment < ApplicationRecord
   # make another old_display_filename see above
   def display_filename
     filename = self.filename
-    unless incoming_message.nil?
+    unless locked? || incoming_message.nil?
       filename = info_request.apply_censor_rules_to_text(filename)
     end
     # Sometimes filenames have e.g. %20 in - no point butchering that
@@ -291,10 +282,20 @@ class FoiAttachment < ApplicationRecord
   def update_and_log_event(event: {}, **params)
     return false unless update(params)
 
+    replaced = locked? && (
+               replacement_body_previously_changed? ||
+               replacement_file_previously_changed?)
+
     log_event(
       'edit_attachment',
       event.merge(
         attachment_id: id,
+        old_locked: locked_previously_was,
+        locked: locked,
+        replaced: replaced,
+        replaced_at: replaced ? replaced_at : nil,
+        replaced_filename: replaced ? filename : nil,
+        replaced_reason: replaced ? replaced_reason : nil,
         old_prominence: prominence_previously_was,
         prominence: prominence,
         old_prominence_reason: prominence_reason_previously_was,
@@ -303,7 +304,67 @@ class FoiAttachment < ApplicationRecord
     )
   end
 
+  def locking?
+    locked? && locked_changed?
+  end
+
+  def unlocking?
+    !locked? && locked_changed?
+  end
+
+  def replacing?
+    !unlocking? && (replacement_file_changed? || replacement_body_changed?)
+  end
+
+  def replaced?
+    replaced_at.present?
+  end
+
+  def replacing_or_replaced?
+    replacing? || replaced?
+  end
+
+  def replaced_filename
+    return filename if replaced? && !replaced_filename_changed?
+
+    super
+  end
+
+  def replacement_body
+    super || body
+  end
+
+  def replacement_body=(new_replacement_body)
+    super unless normalize_line_endings(new_replacement_body) ==
+                 normalize_line_endings(body)
+  end
+
   private
+
+  def mail_attributes
+    MailHandler.attachment_attributes_for_hexdigest(
+      raw_email.mail,
+      hexdigest: hexdigest
+    )
+
+  rescue MailHandler::MismatchedAttachmentHexdigest
+    begin
+      attributes = MailHandler.attempt_to_find_original_attachment_attributes(
+        raw_email.mail,
+        body: file.download
+      ) if file.attached?
+
+    rescue ActiveStorage::FileNotFoundError
+      raise MissingAttachment, "attachment missing from storage (ID=#{id})"
+    end
+
+    unless attributes
+      raise MissingAttachment, "attachment missing in raw email (ID=#{id})"
+    end
+
+    update(hexdigest: attributes[:hexdigest])
+    attributes
+  end
 
   def load_attachment_from_incoming_message!
     attachment = load_attachment_from_incoming_message
@@ -315,5 +376,55 @@ class FoiAttachment < ApplicationRecord
 
   def text_type?
     AlaveteliTextMasker::TextMask.include?(content_type)
+  end
+
+  def handle_locked
+    if unlocking? && replaced?
+      file_blob.upload(StringIO.new(unmasked_body), identify: false)
+      file_blob.save
+
+      self.replaced_at = nil
+      self.replaced_reason = nil
+    end
+
+    if unlocking?
+      self.masked_at = nil
+      self.filename = mail_attributes[:filename]
+      ensure_filename!
+    end
+
+    if locking? || unlocking?
+      FoiAttachmentMaskJob.perform_later(self) unless masked_at
+    end
+
+    true
+  end
+
+  def handle_replacements
+    if replacing? || (replaced? && replaced_filename_changed?)
+      self.filename = replaced_filename.presence ||
+                      replacement_file&.original_filename ||
+                      mail_attributes[:filename]
+      ensure_filename!
+    end
+
+    if replacing?
+      self.replaced_at = Time.zone.now
+      self.masked_at = Time.zone.now
+      self.locked = true
+
+      if replacement_file_changed?
+        file.attach(
+          io: replacement_file,
+          filename: filename,
+          content_type: content_type
+        )
+      elsif replacement_body_changed?
+        file_blob.upload(StringIO.new(replacement_body), identify: false)
+        file_blob.save
+      end
+    end
+
+    true
   end
 end