+++ doc/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Highlighted Features
 
-* Add attachment locking (Graeme Porteous)
+* Add attachment locking and replacement (Graeme Porteous)
 * Only send clarification notifications to recent requests (Gareth Rees)
 * Improve citation form field width (Gareth Rees)
 * Add report link to user profile pages (Gareth Rees)